separate properly the two cpu implementations
