+++ include/instructions.h
@@ -1,0 +1,9 @@
+#ifndef INCLUDE_INSTRUCTIONS_H_
+#define INCLUDE_INSTRUCTIONS_H_
+
+#include "cpu_op.h"
+
+extern const struct s_cpu_z80 instructions_cb[];
+extern const struct s_cpu_z80 instructions_base[];
+
+#endif /* INCLUDE_INSTRUCTIONS_H_ */
